requirements.txt
aniso8601==1.3.0
beautifulsoup4==4.6.0
bkcharts==0.2
bleach==2.0.0
bokeh==0.12.7
branca==0.2.0
certifi==2017.11.5
chardet==3.0.4
click==6.7
colorama==0.3.9
cycler==0.10.0
DateTime==4.2
decorator==4.1.2
dj-database-url==0.4.2
Django==1.11.7
django-bootstrap-datepicker==1.2.3
django-bootstrap-form==3.3
django-crispy-forms==1.7.0
django-otp==0.4.2
django-registration-redux==1.8
entrypoints==0.2.3
Flask==0.12.2
Flask-JWT==0.3.2
Flask-RESTful==0.3.6
Flask-SQLAlchemy==2.3.2
folium==0.3.0
future==0.16.0
geopy==1.11.0
gunicorn==19.7.1
html5lib==0.999999999
idna==2.6
ipykernel==4.6.1
ipython==6.1.0
ipython-genutils==0.2.0
ipywidgets==6.0.0
itsdangerous==0.24
jedi==0.10.2
Jinja2==2.9.6
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.1.0
jupyter-console==5.1.0
jupyter-core==4.3.0
lxml==4.1.1
MarkupSafe==1.0
matplotlib==2.0.2
mistune==0.7.4
nbconvert==5.2.1
nbformat==4.3.0
networkx==2.0
notebook==5.0.0
numpy==1.13.1
olefile==0.44
opencv-python==3.4.0.12
pandas==0.20.2
pandocfilters==1.4.1
pickleshare==0.7.4
Pillow==4.3.0
prompt-toolkit==1.0.14
psycopg2==2.7.3.2
Pygments==2.2.0
PyInstaller==3.2.1
PyJWT==1.4.2
pyparsing==2.2.0
pypiwin32==220
python-dateutil==2.6.0
pytz==2017.2
PyYAML==3.12
pyzmq==16.0.2
qrcode==5.3
qtconsole==4.3.0
requests==2.18.4
scikit-learn==0.19.0
simplegeneric==0.8.1
six==1.10.0
SQLAlchemy==1.1.15
testpath==0.3.1
tornado==4.5.1
traitlets==4.3.2
urllib3==1.22
virtualenv==15.1.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.12.2
widgetsnbextension==2.0.0
zope.interface==4.4.3
